p384: refactor `MODULUS` and macros
- Makes `MODULUS` a toplevel constant instead of an inherent one, and
  also makes it a `U384` instead of a `FieldElement`, since it isn't
  technically a valid field element.
- Factors arithmetic macros into their own module in preparation for
  using them to impl `Scalar`.
